# Optimize DatabaseProjectStoreManager.getProjectsWithLatestRevision