--- a/digdag-core/src/main/java/io/digdag/core/database/DatabaseProjectStoreManager.java
+++ b/digdag-core/src/main/java/io/digdag/core/database/DatabaseProjectStoreManager.java
@@ -516,22 +516,26 @@ public class DatabaseProjectStoreManager
     public interface H2Dao
             extends Dao
     {
-        // h2 doesn't have max window function.
-        // this is not efficient but gave up optimization on h2.
+        // H2 doesn't support MAX window function or LATERAL.
         @Override
         @SqlQuery("select proj.*, rev.name as revision_name, rev.created_at as revision_created_at, rev.archive_type as revision_archive_type, rev.archive_md5 as revision_archive_md5" +
-        " from projects proj" +
-        " join revisions rev on proj.id = rev.project_id" +
-        " join (" +
-            "select project_id, max(id) AS id" +
-            " from revisions" +
+        " from (" +
+            "select project_id, max(rev.id) as latest_revision_id" +
+            " from revisions rev" +
+            " join (" +
+                "select id" +
+                " from projects proj" +
+                " where proj.site_id = :siteId" +
+                " and proj.name is not null" +
+                " and proj.id > :lastId" +
+                " order by id asc" +
+                " limit :limit" +
+            ") match_proj on rev.project_id = match_proj.id" +
             " group by project_id" +
-        ") a on a.id = rev.id" +
-        " where proj.site_id = :siteId" +
-        " and proj.name is not null" +
-        " and proj.id > :lastId" +
-        " order by proj.id asc" +
-        " limit :limit")
+        ") proj_rev" +
+        " join projects proj on proj.id = proj_rev.project_id" +
+        " join revisions rev on rev.id = proj_rev.latest_revision_id" +
+        " order by id asc")
         List<StoredProjectWithRevision> getProjectsWithLatestRevision(@Bind("siteId") int siteId, @Bind("limit") int limit, @Bind("lastId") int lastId);
 
         // h2's MERGE doesn't return generated id when conflicting row already exists
@@ -569,16 +573,16 @@ public class DatabaseProjectStoreManager
             extends Dao
     {
         @Override
-        @SqlQuery("select id, site_id, name, created_at, deleted_at, deleted_name, revision_name, revision_created_at, revision_archive_type, revision_archive_md5" +
-        " from (" +
-            " select proj.*, rev.name as revision_name, rev.created_at as revision_created_at, rev.archive_type as revision_archive_type, rev.archive_md5 as revision_archive_md5, rev.id as revision_id, max(rev.id) OVER(partition by rev.project_id) as max_revision_id" +
-            " from projects proj" +
-            " join revisions rev on proj.id = rev.project_id" +
-            " where proj.site_id = :siteId" +
-            " and proj.name is not null" +
-            " and proj.id > :lastId" +
-        ") as projects_with_revision" +
-        " where projects_with_revision.revision_id = projects_with_revision.max_revision_id" +
+        @SqlQuery("select id, site_id, name, created_at, deleted_at, deleted_name, rev.*" +
+        " from projects proj, lateral (" +
+            "select rev.name as revision_name, rev.created_at as revision_created_at, rev.archive_type as revision_archive_type, rev.archive_md5 as revision_archive_md5" +
+            " from revisions rev" +
+            " where rev.project_id = proj.id" +
+            " order by rev.id desc" +
+            " limit 1" +
+        ") rev" +
+        " where site_id = :siteId" +
+        " and id > :lastId" +
         " order by id asc" +
         " limit :limit")
         List<StoredProjectWithRevision> getProjectsWithLatestRevision(@Bind("siteId") int siteId, @Bind("limit") int limit, @Bind("lastId") int lastId);